fix(app): check do treino de volta na tela, reset semanal e quebras em 4 larguras
Ordem do dono: "tem celular de vários tamanhos, tem que atualizar pra todos".

A FERRAMENTA ESTAVA CEGA. A varredura rodava só a 430px — o aparelho do dono
tem 360. Por isso ela vinha "19 telas limpas" enquanto ele via a caixa de
seleção fora da margem. Agora roda em 320/360/390/412 (320 é o piso real:
aparelho pequeno ou fonte grande do sistema) e aprendeu duas coisas:
distinguir faixa que rola de propósito de vazamento real (senão o número sai
inflado — a 1ª rodada acusou 53, sendo a maioria aba rolável normal), e
enxergar CONTROLE SEM TEXTO, que é como o ✅ do treino escapava.

TREINO — O ✅ SUMIA. A linha é uma grade de 4 colunas e a do nome não tinha
min-w-0: célula 1fr não encolhe abaixo do conteúdo, então exercício de nome
comprido esticava a grade e empurrava a 4ª coluna pra fora da tela. Sumia o
CONTROLE, não o texto — não dava pra marcar o treino. Agora a coluna trunca e
o check fica ancorado.

TREINO — NUNCA REINICIAVA. "Fiz esse treino terça passada mas ainda estava
marcado como feito hoje." A causa é a forma do dado: `done` mora dentro do
plano, na chave do DIA DA SEMANA, sem data nenhuma — marcar terça grava
"TERÇA feito" pra sempre. Em vez de pendurar data em cada exercício (mudaria
o formato salvo de todo mundo), guardo a QUAL SEMANA os checks pertencem;
semana nova, checks zerados, uma vez só. Semana começa na segunda, que é o
desenho do módulo.

REVER TUTORIAL — A SAÍDA ERA INVISÍVEL. Ela existia: um ✕ cinza de 36px sem
rótulo, no canto. A tela seguinte tinha um botão largo escrito "Encerrar
tutorial". Duas saídas diferentes pra mesma ação, e a de cá era a que ninguém
achava. Agora é a mesma nas duas.

QUEBRAS: selo "19 dias consecutivos" partia em duas linhas (faltava trava);
saudação virava "Boa tarde, João Vic…" (agora só o primeiro nome, e fonte
menor a 320px — as iniciais do avatar seguem com o nome inteiro); resumo do
Finanças eram 4 colunas de valores grandes numa faixa rolável que parecia
cortada (virou 2×2, nada rola); Beleza vazava 31px por falta de min-w-0; data
do diário encurtada pra caber.

BEM-VINDO DO TUTORIAL: era logo em escala de cinza e botão preto puro, a única
tela do fluxo sem cor, num app cuja marca é grafite + magenta. Agora fala a
mesma língua da comemoração pós-compra.

MEDIDO: vazamentos 7 → 0 nas quatro larguras (total de apontamentos 46 → 33; o
resto que sobra é ruído do detector — "JO" é o avatar redondo, o de planos é
parágrafo de duas linhas mesmo). No APK: nenhum controle fora da tela em
treino, home e finanças. Auditoria do tutorial e smoke dos funis verdes.

versionCode 28 / 1.0.27

+++ src/pages/DesenvolvimentoPessoal.tsx
@@ -371,7 +371,7 @@ const DesenvolvimentoPessoal = () => {
               <div className="space-y-2">
                 {Object.entries(journalEntries).filter(([d]) => d !== today).sort((a, b) => b[0].localeCompare(a[0])).slice(0, 10).map(([date, entry]) => (
                   <div key={date} className="bg-muted/30 rounded-lg p-3 border border-border">
-                    <p className="text-xs font-bold mb-1">{new Date(date + "T12:00:00").toLocaleDateString("pt-BR", { weekday: "long", day: "numeric", month: "long" })}</p>
+                    <p className="text-xs font-bold mb-1">{new Date(date + "T12:00:00").toLocaleDateString("pt-BR", { weekday: "short", day: "numeric", month: "long" })}</p>
                     <p className="text-[10px] text-muted-foreground italic mb-1">Prompt: {entry.prompt}</p>
                     <p className="text-xs">{entry.text}</p>
                   </div>
@@ -490,7 +490,7 @@ const DesenvolvimentoPessoal = () => {
               <h3 className="text-xs font-bold mb-3">📅 HISTÓRICO</h3>
               {Object.entries(gratitudeEntries).sort((a, b) => b[0].localeCompare(a[0])).slice(0, 14).map(([date, items]) => (
                 <div key={date} className="bg-muted/30 rounded-lg p-3 border border-border mb-2">
-                  <p className="text-xs font-bold mb-1">{new Date(date + "T12:00:00").toLocaleDateString("pt-BR", { weekday: "long", day: "numeric", month: "long" })}</p>
+                  <p className="text-xs font-bold mb-1">{new Date(date + "T12:00:00").toLocaleDateString("pt-BR", { weekday: "short", day: "numeric", month: "long" })}</p>
                   {(Array.isArray(items) ? items : []).filter(Boolean).map((item, i) => <p key={i} className="text-xs text-muted-foreground">• {item}</p>)}
                 </div>
               ))}
